Choice: Fix class inheritance
