+++ src/euphorie/client/browser/choice.py
@@ -1,17 +1,20 @@
-# FIXME fix inheritance
-from euphorie.client.browser.risk import IdentificationView as RiskIdentificationView
 from plone import api
 from plone.memoize.instance import memoize
 from Products.Five import BrowserView
 
 
-class IdentificationView(RiskIdentificationView):
+class IdentificationView(BrowserView):
     """A view for displaying a choice in the identification phase."""
 
     @property
     @memoize
     def webhelpers(self):
         return api.content.get_view("webhelpers", self.context.aq_parent, self.request)
+
+    @property
+    @memoize
+    def navigation(self):
+        return api.content.get_view("navigation", self.context, self.request)
 
     @property
     @memoize
@@ -25,24 +28,16 @@ class IdentificationView(RiskIdentificationView):
     def selected(self):
         return [option.zodb_path for option in self.context.options]
 
-    @property
-    @memoize
-    def risk(self):
-        # XXX don't inherit this method
-        return self.context.aq_parent.aq_parent.restrictedTraverse(
-            self.context.zodb_path.split("/")
-        )
-
-    @property
-    def is_custom_risk(self):
-        # XXX don't inherit this method
-        return False
-
     def set_answer_data(self, reply):
         answer = reply.get("answer", [])
         if not isinstance(answer, (list, tuple)):
             answer = [answer]
         # XXX Check if paths are valid?
+        # for path in answer[:]:
+        #     try:
+        #         self.webhelpers.traversed_session.restrictedTraverse(path)
+        #     except KeyError:
+        #        answer.remove(path)
         self.context.set_options_by_zodb_path(answer)
 
     def __call__(self):
@@ -55,6 +50,5 @@ class IdentificationView(RiskIdentificationView):
         if self.request.method == "POST":
             reply = self.request.form
             self.set_answer_data(reply)
-            return self.proceed_to_next(reply)
-        else:
-            return self.index()
+            return self.navigation.proceed_to_next(reply)
+        return self.index()
